// src/lib/def/api/v1/info.ts
export type InfoAll = {
  ProjectsCount: number | string;
  VersionsCount: number | string;
  AssetsCount: number | string;
  TypesCount: number | string;
  UsersCount: number | string;
  UpdateCheckCount: number | string;
  UpdatePushCount: number | string;
  HttpReqCount: number | string;
  UpdateCheckCountDaily: number | string;
  UpdatePushCountDaily: number | string;
  HttpReqCountDaily: number | string;
  UserProjectCount: number | string;
};
export type InfoProj = {
  ProjectsCount: number | string;
  UserProjectCount: number | string;
};